feat : 초기설정
