--- jigeum-api/build.gradle
@@ -1,26 +1,0 @@
-tasks.named('bootJar'){
-    enabled = true
-}
-
-tasks.named('jar'){
-    enabled = false
-}
-
-bootJar {
-    enabled = true
-    mainClass = 'se.sowl.jigeumapi.JigeumApiApplication'
-}
-
-jar {
-    enabled = true
-}
-dependencies {
-    implementation project(':jigeum-domain')
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.springframework.boot:spring-boot-starter-webflux'
-    implementation 'org.springframework.boot:spring-boot-starter-security'
-    implementation 'org.springframework.security:spring-security-test'
-    implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-}
