--- jigeum-api/src/main/java/se/sowl/jigeumapi/common/config/CacheConfig.java
@@ -1,13 +1,0 @@
-package se.sowl.jigeumapi.common.config;
-
-// TODO: if use cache, active this configuration
-//@EnableCaching
-//@Configuration
-//public class CacheConfig {
-//    @Bean
-//    public CacheManager cacheManager() {
-//        ConcurrentMapCacheManager cacheManager = new ConcurrentMapCacheManager();
-//        cacheManager.setCacheNames(Arrays.asList("crowdnessMap", "gyms"));
-//        return cacheManager;
-//    }
-//}
