--- jigeum-api/src/main/java/se/sowl/jigeumapi/common/config/HttpClientConfig.java
@@ -1,25 +1,0 @@
-package se.sowl.jigeumapi.common.config;
-
-import io.netty.channel.ChannelOption;
-import io.netty.handler.timeout.ReadTimeoutHandler;
-import io.netty.handler.timeout.WriteTimeoutHandler;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import reactor.netty.http.client.HttpClient;
-
-import java.time.Duration;
-import java.util.concurrent.TimeUnit;
-
-@Configuration
-public class HttpClientConfig {
-
-    @Bean
-    public HttpClient httpClient() {
-        return HttpClient.create()
-                .option(ChannelOption.CONNECT_TIMEOUT_MILLIS, 6000)
-                .responseTimeout(Duration.ofMillis(6000))
-                .doOnConnected(conn ->
-                        conn.addHandlerLast(new ReadTimeoutHandler(5000, TimeUnit.MILLISECONDS))
-                            .addHandlerLast(new WriteTimeoutHandler(5000, TimeUnit.MILLISECONDS)));
-    }
-}
