--- jigeum-api/src/main/java/se/sowl/jigeumapi/common/config/JpaConfig.java
@@ -1,9 +1,0 @@
-package se.sowl.jigeumapi.common.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-@Configuration
-@EnableJpaRepositories(basePackages = "se.sowl.jigeumdomain")
-public class JpaConfig {
-}
