--- jigeum-api/src/main/java/se/sowl/jigeumapi/common/config/OAuth2AuthenticationSuccessHandler.java
@@ -1,26 +1,0 @@
-package se.sowl.jigeumapi.common.config;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import jakarta.servlet.http.HttpSession;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Component
-public class OAuth2AuthenticationSuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
-    @Value("${spring.front.url}")
-    private String frontUrl;
-
-    @Override
-    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response,
-                                        Authentication authentication) throws IOException, ServletException {
-        HttpSession session = request.getSession(true);
-         session.setAttribute("user", authentication.getPrincipal());
-        getRedirectStrategy().sendRedirect(request, response, frontUrl);
-    }
-}
