--- jigeum-api/src/main/java/se/sowl/jigeumapi/user/service/UserService.java
@@ -1,18 +1,0 @@
-package se.sowl.jigeumapi.user.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import se.sowl.jigeumdomain.user.domain.User;
-import se.sowl.jigeumdomain.user.repository.UserRepository;
-
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-public class UserService {
-    private final UserRepository userRepository;
-
-    public List<User> getList() {
-        return userRepository.findAll();
-    }
-}
