--- jigeum-api/src/test/java/se/sowl/jigeumapi/JigeumApiApplicationTests.java
@@ -1,13 +1,0 @@
-package se.sowl.jigeumapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class JigeumApiApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
