--- jigeum-domain/src/main/java/se/sowl/jigeumdomain/oauth/domain/OAuth2Provider.java
@@ -1,14 +1,0 @@
-package se.sowl.jigeumdomain.oauth.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum OAuth2Provider {
-    GOOGLE("google"),
-    NAVER("naver"),
-    KAKAO("kakao");
-
-    private final String registrationId;
-}
