--- jigeum-domain/src/main/java/se/sowl/jigeumdomain/oauth/domain/Role.java
@@ -1,12 +1,0 @@
-package se.sowl.jigeumdomain.oauth.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum Role {
-    USER("USER");
-
-    private final String value;
-}
