--- jigeum-domain/src/main/java/se/sowl/jigeumdomain/user/domain/User.java
@@ -1,46 +1,0 @@
-package se.sowl.jigeumdomain.user.domain;
-
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.time.LocalDateTime;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Entity
-@Table(name = "users")
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-    private String nickname;
-
-    @Column(unique = true, nullable = false)
-    private String email;
-
-    @Column(nullable = false)
-    private String provider;
-
-    @CreationTimestamp
-    @Column(name = "created_at", updatable = false)
-    private LocalDateTime createdAt;
-
-    @Builder
-    public User(Long id, String name, String nickname, String email, String provider) {
-        this.id = id;
-        this.name = name;
-        this.nickname = nickname;
-        this.email = email;
-        this.provider = provider;
-    }
-}
