--- jigeum-domain/src/main/resources/application.yml
@@ -1,46 +1,0 @@
-spring:
-  profiles:
-    group:
-      test: test
-      prod: prod
----
-spring:
-  config:
-    activate:
-      on-profile: test
-  h2:
-    console:
-      enabled: true
-  datasource:
-    url: jdbc:h2:mem:jigeum-local;MODE=MySQL
-    username: user
-    password:
-    driver-class-name: org.h2.Driver
-  jpa:
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.MySQL8Dialect
-        show_sql: true
-    hibernate:
-      ddl-auto: create
----
-spring:
-  config:
-    activate:
-      on-profile: prod
-  h2:
-    console:
-      enabled: true
-  datasource:
-    url: jdbc:mysql://localhost:3306/jigeum?useUnicode=yes&characterEncoding=UTF-8&allowMultiQueries=true&serverTimezone=Asia/Seoul
-    username:
-    password:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-  jpa:
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.MySQL8Dialect
-        show_sql: true
-    hibernate:
-      ddl-auto: none
----
