+++ postHubing-api/build.gradle
@@ -1,0 +1,26 @@
+tasks.named('bootJar'){
+    enabled = true
+}
+
+tasks.named('jar'){
+    enabled = false
+}
+
+bootJar {
+    enabled = true
+    mainClass = 'se.sowl.postHubingapi.PostHubingApiApplication'
+}
+
+jar {
+    enabled = true
+}
+dependencies {
+    implementation project(':postHubing-domain')
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    implementation 'org.springframework.boot:spring-boot-starter-webflux'
+    implementation 'org.springframework.boot:spring-boot-starter-security'
+    implementation 'org.springframework.security:spring-security-test'
+    implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+}
