+++ postHubing-api/src/main/java/se/sowl/postHubingapi/PostHubingApiApplication.java
@@ -1,0 +1,17 @@
+package se.sowl.postHubingapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.cache.annotation.EnableCaching;
+
+@SpringBootApplication
+@EntityScan(basePackages = {"se.sowl.postHubingdomain"})
+@EnableCaching
+public class PostHubingApiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(PostHubingApiApplication.class, args);
+    }
+
+}
