+++ postHubing-api/src/main/java/se/sowl/postHubingapi/common/config/CacheConfig.java
@@ -1,0 +1,13 @@
+package se.sowl.postHubingapi.common.config;
+
+// TODO: if use cache, active this configuration
+//@EnableCaching
+//@Configuration
+//public class CacheConfig {
+//    @Bean
+//    public CacheManager cacheManager() {
+//        ConcurrentMapCacheManager cacheManager = new ConcurrentMapCacheManager();
+//        cacheManager.setCacheNames(Arrays.asList("crowdnessMap", "gyms"));
+//        return cacheManager;
+//    }
+//}
