+++ postHubing-api/src/main/java/se/sowl/postHubingapi/common/config/JpaConfig.java
@@ -1,0 +1,9 @@
+package se.sowl.postHubingapi.common.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@Configuration
+@EnableJpaRepositories(basePackages = "se.sowl.postHubingdomain")
+public class JpaConfig {
+}
