+++ postHubing-api/src/main/java/se/sowl/postHubingapi/common/config/OAuth2AuthenticationSuccessHandler.java
@@ -1,0 +1,29 @@
+package se.sowl.postHubingapi.common.config;
+
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
+import org.springframework.stereotype.Component;
+
+import java.io.IOException;
+
+@Component
+public class OAuth2AuthenticationSuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
+
+    // TODO: targetURL 환경변수화 처리
+    @Override
+    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response,
+                                        Authentication authentication) throws IOException, ServletException {
+        HttpSession session = request.getSession(true);
+        session.setAttribute("user", authentication.getPrincipal());
+        String targetUrl = "http://localhost:3000/";
+        getRedirectStrategy().sendRedirect(request, response, targetUrl);
+
+    }
+}
+
+
