+++ postHubing-api/src/main/java/se/sowl/postHubingapi/common/config/OAuth2Config.java
@@ -1,0 +1,14 @@
+package se.sowl.postHubingapi.common.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
+
+@Configuration
+public class OAuth2Config {
+
+    @Bean
+    public DefaultOAuth2UserService defaultOAuth2UserService() {
+        return new DefaultOAuth2UserService();
+    }
+}
