+++ postHubing-api/src/main/java/se/sowl/postHubingapi/user/controller/UserController.java
@@ -1,0 +1,23 @@
+package se.sowl.postHubingapi.user.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import se.sowl.postHubingapi.common.CommonResponse;
+
+@RestController
+@RequestMapping("/api/users")
+@RequiredArgsConstructor
+public class UserController {
+
+    @GetMapping("/me")
+    @PreAuthorize("isAuthenticated()")
+    public CommonResponse<Void> getMe(@AuthenticationPrincipal OAuth2User user) {
+        System.out.println(user.toString());
+        return CommonResponse.ok();
+    }
+}
