+++ postHubing-api/src/main/java/se/sowl/postHubingapi/user/service/UserService.java
@@ -1,0 +1,18 @@
+package se.sowl.postHubingapi.user.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import se.sowl.postHubingdomain.user.domain.User;
+import se.sowl.postHubingdomain.user.repository.UserRepository;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class UserService {
+    private final UserRepository userRepository;
+
+    public List<User> getList() {
+        return userRepository.findAll();
+    }
+}
