+++ postHubing-api/src/test/java/se/sowl/postHubingapi/PostHubingApiApplicationTests.java
@@ -1,0 +1,13 @@
+package se.sowl.postHubingapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PostHubingApiApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
