+++ postHubing-domain/src/main/java/se/sowl/postHubingdomain/oauth/domain/OAuth2Provider.java
@@ -1,0 +1,14 @@
+package se.sowl.postHubingdomain.oauth.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum OAuth2Provider {
+    GOOGLE("google"),
+    NAVER("naver"),
+    KAKAO("kakao");
+
+    private final String registrationId;
+}
