+++ postHubing-domain/src/main/java/se/sowl/postHubingdomain/oauth/domain/Role.java
@@ -1,0 +1,12 @@
+package se.sowl.postHubingdomain.oauth.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum Role {
+    USER("USER");
+
+    private final String value;
+}
