+++ postHubing-domain/src/main/java/se/sowl/postHubingdomain/user/domain/CustomOAuth2User.java
@@ -1,0 +1,33 @@
+package se.sowl.postHubingdomain.user.domain;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+
+import java.util.Collection;
+import java.util.Map;
+
+@RequiredArgsConstructor
+public class CustomOAuth2User implements OAuth2User {
+    private final User user;
+    private final OAuth2User oauth2User;
+
+    @Override
+    public Map<String, Object> getAttributes() {
+        return oauth2User.getAttributes();
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return oauth2User.getAuthorities();
+    }
+
+    @Override
+    public String getName() {
+        return oauth2User.getName();
+    }
+
+    public Long getUserId() {
+        return user.getId();
+    }
+}
