+++ postHubing-domain/src/main/java/se/sowl/postHubingdomain/user/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package se.sowl.postHubingdomain.user.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import se.sowl.postHubingdomain.user.domain.User;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmailAndProvider(String email, String provider);
+}
