+++ settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'jigeum'
+rootProject.name = 'postHubing'
 
-include 'jigeum-api', 'jigeum-domain'
+include 'postHubing-api', 'postHubing-domain'
